+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Resizable Panes - Persistence. Refs UIOR-683.
 * Apply Results list column chooser. Refs UIOR-691.
 * Filter order by subscription = yes/no. Refs UIOR-710.
+* Add linked instances accordion to the POL. Refs UIOR-689.
 
 ## [2.3.1](https://github.com/folio-org/ui-orders/tree/v2.3.1) (2021-04-16)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v2.3.0...v2.3.1)